issue-draft-creator/pyproject.toml
[project]
name = "issue-draft-creator"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "google-generativeai (>=0.3.0,<0.4.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
